feat(tizen): Migrate Google Container Registry to GitHub Container Registry
Google Container Registry is imposing more painful restrictions on us,
so we are migrating the Tizen Studio TV docker image from GCR to GHCR
(GitHub Container Registry).  This should be simpler overall for an
open source project hosted on GitHub, and should also work
transparently for forks.

This also adds a new workflow to automate the publication of those
docker images.